+++ {{cookiecutter.github_repo_name}}/terraform/environments/modules/kubernetes_ingress_clb/variables.tf
@@ -17,13 +17,14 @@ variable "shared_resource_namespace" {
   type = string
 }
 
-variable "namespace" {
-  type = string
-}
 
 variable "root_domain" {
   description = "Root domain (route53 zone) for the default cluster ingress."
   type        = string
+}
+
+variable "aws_region" {
+  type = string
 }
 
 variable "tags" {