+++ {{cookiecutter.github_repo_name}}/terraform/stacks/modules/kubernetes/addon_ebs_csi_driver.tf
@@ -27,7 +27,7 @@ resource "aws_iam_role" "AmazonEKS_EBS_CSI_DriverRole" {
       {
         "Effect" : "Allow",
         "Principal" : {
-          "Federated" : "arn:aws:iam::293205054626:oidc-provider/${module.eks.oidc_provider}"
+          "Federated" : "arn:aws:iam::${var.account_id}:oidc-provider/${module.eks.oidc_provider}"
         },
         "Action" : "sts:AssumeRoleWithWebIdentity",
         "Condition" : {
@@ -55,7 +55,6 @@ resource "null_resource" "annotate-ebs-csi-controller" {
   provisioner "local-exec" {
     command = <<-EOT
       kubectl annotate serviceaccount ebs-csi-controller-sa -n kube-system eks.amazonaws.com/role-arn=arn:aws:iam::${var.account_id}:role/${aws_iam_role.AmazonEKS_EBS_CSI_DriverRole.name}
-      kubectl rollout restart deployment ebs-csi-controller -n kube-system
       kubectl rollout restart deployment ebs-csi-controller -n kube-system
     EOT
   }