+++ {{cookiecutter.github_repo_name}}/terraform/stacks/modules/kubernetes_kubeapps/variables.tf
@@ -1,0 +1,23 @@
+#------------------------------------------------------------------------------
+# written by: Lawrence McDaniel
+#             https://lawrencemcdaniel.com/
+#
+# date: Mar-2022
+#
+# usage: create an EKS cluster
+#------------------------------------------------------------------------------
+
+variable "admin_domain" {
+  type = string
+
+}
+
+variable "stack_namespace" {
+  type = string
+}
+
+variable "tags" {
+  description = "A map of tags to add to all resources. Tags added to launch configuration or templates override these values for ASG Tags only."
+  type        = map(string)
+  default     = {}
+}