+++ {{cookiecutter.github_repo_name}}/terraform/stacks/modules/kubernetes_monitoring/install_metrics-server.tf
@@ -1,0 +1,44 @@
+#------------------------------------------------------------------------------
+#
+# see: https://github.com/kubernetes-sigs/metrics-server
+#
+#------------------------------------------------------------------------------
+
+#-----------------------------------------------------------
+# written by: Lawrence McDaniel
+#             https://lawrencemcdaniel.com/
+#
+# date: Aug-2022
+#
+# usage: installs Kubernetes metrics-server.
+#
+# NOTE: you must initialize a local helm repo in order to run
+# this script.
+#
+#   brew install helm
+#   helm repo add metrics-server https://kubernetes-sigs.github.io/metrics-server/
+#   helm repo update
+#   helm search repo metrics-server
+#
+#-----------------------------------------------------------
+
+
+resource "helm_release" "metrics_server" {
+  namespace        = "monitoring"
+  create_namespace = true
+
+  name       = "metrics-server"
+  repository = "https://kubernetes-sigs.github.io/metrics-server/"
+  chart      = "metrics-server"
+  version    = "~> 3.8"
+
+}
+
+resource "kubectl_manifest" "vpa-metrics-server" {
+  yaml_body = file("${path.module}/yml/verticalpodautoscalers/vpa-metrics-server.yaml")
+
+  depends_on = [
+    helm_release.vpa,
+    helm_release.metrics_server
+  ]
+}